fix: exit with correct status code (#8)
Fixes #7

Notes:
- Added `@types/node` dev dependencies to improve type safety, notably
for `process` for example, and also verifying that the user gives a URL
as third argument. It can be out of scope of this PR, so feel free to
tell me if you want to add that in a separated PR.
- Added a new test case in `test/cli-test.js`, but it only has the
"happy-path" scenario where all links are valid. I'm not sure what
website, we could take for the test, to verify that it exits with status
code 1 if there are broken links.
- ESLint, Prettier and editorconfig are not configured on the project, I
tried to follow the project conventions with the code already available,
forgive me if I forgot a `;` or semicolon or something. :sweat_smile:
However, it could be great to integrate these tools to improve
consistency (could be done in a separate PR).

Co-authored-by: Steven <steven@ceriously.com>